Rename & move gRPC-specific timeout code
